Show remaining time when scrubbing
This commit introduces a property in ZpoolScrub class which enables us to get remaining seconds left for an active scrub. It also adds a condition where we return None for end time of a scrub if we have an active scrub in progress as there is no end time at that point and the time returned is start of the epoch.
Ticket: #56475