Improve snapshots retrieval speed
This commit introduces a change to make sure that we are able to retrieve snapshots faster then the current method we have for it. With the number of snapshots going higher, the time being taken to retrieve their properties gets even higher. Changes introduces here make sure that the new speed is on level with the zfs cli.
Ticket: #83580