Cython str references bug fix
This commit fixes a case where if the original reference to python str isn't maintained, the c string can be unpredictable based on when when it is garbage collected or a new assignment is made in which case the old c str can point to the new assignment. The issue got detected via attach method which represented this behaviour. However for good measure, some other usages have been covered as well to take this factor into account.