Allow setting props in top level dict
This commit updates serialized datasets api to allow setting different props to be in the top level dict when the data is retrieved. One thing to note here is that if a key is specified to be in the top level dict and hasn't been specified by props attribute, it will raise a value error and also keeping in line with legacy implementation, we add mountpoint key by default.