Do not treat mountpoint specially
Recently we removed usage(s) of having top level props in e606cd583951e5836da47d7d379a6c523e790f1b. We should remove the only property which is being added to top level in the dict which is mountpoint for 2 reasons - first being that right now it's dependent on mounted property which might not be retrieved if not explicitly specified, secondly this is internal API and in pool.dataset.query, there shouldn't be any motivation to treat this specially.