devel/py-libzfs: fix scrub/resilver percentage and time remaining
The percentage calculation was using bytes issued per pass instead of total
bytes issued, and as a result it was wrong after a reboot.

The total_secs_left calculation was using the bytes examined, not the bytes
issued, so it was underestimating the amount of work remaining, and could
even go negative.