Initialize csource before zfs_prop_get
As of openzfs/zfs@e4288a839 the src passed to zfs_prop_get() must be
initialized by the caller.  It can be set to ZPROP_SRC_NONE to preserve
the existing behavior and get properties with any source, or it can be
set to a source to filter by.

Initialize csource to ZPROP_SRC_NONE before calls to zfs_prop_get() to
preserve the existing behavior.