Ignore errors while setting sharesmb/sharenfs properties
From UI, pool import fails for users that have sharesmb and/or
sharenfs properties set for dataset/s. This failure occurs because
ZFSDatasetService fails to update properties of the new imported
dataset/s.

ZFS provided sharing mechanism is not compatible with how it is
implemented in TrueNAS, so when these properties are set, failure
occurs because sharing is not successful.

These failures while updating the properties should be ignored,
and pool import should be allowed to continue, with sharesmb and/or
sharenfs set.

Signed-off-by: Umer Saleem <usaleem@ixsystems.com>